requirements.txt
Flask==3.1.1
groq==0.30.0
numpy>=1.26.0,<2.2.0
pandas==2.3.1
Pillow==11.3.0
protobuf==4.25.3
pytest==7.4.4
python-dotenv==1.1.1
reportlab==4.4.2
tensorflow==2.19.0
torch==2.7.1
torchvision==0.22.1
transformers==4.53.2
Werkzeug==3.1.3
pymupdf==1.23.22
